assay update perf much improved - using WordSetId pervasively
A remaining hotspot seems to be `idFor(wordSet: BitSet)` as when the
`BitSet` is used as the key in a 'get' operation, once hashcode lookup
is complete, you still need to do a `BitSet.equals()` to check you have
a match - and the resulting `SortedSet.equals` is fairly slow, with
`BitSet.size` dominating.

If we were using optimised 12-bit array-storage, I think we could
have `size` - and even full equality checking - done much quicker.
`BitSet` is using `long` arrays of size 32 to 64 (so ~512 bytes)
to represent a bitfield of 2315 possible words, even though many
of the bitfields are very sparse (~85% of bitfields are >3.7%
populated, and a 12-bit-array representation would take only
128 bytes to represent the same data - and could contain size
in just 12 bits).

https://docs.google.com/spreadsheets/d/1vhRliOSUQnMd3nLfFD6x6rcJiWEJAs02RYi4kzqc9zQ/edit?usp=sharing