Only evaluate if a guess-letter was correct once, rather than twice
Doing this requires introducing a new, additional, custom fold-state
class - where previously the 1st of the 2 passes was simply using a
`FrequencyMap[Letter]` to hold its state, it now uses the
newly-introduced `Census` case class to hold both the frequency-map,
and now also a sequence of `Option[Letter]`, which for each guess-letter
is either `None` if letter was _correct_, or populated with the
guess-letter if the letter was _not_ correct.

This has the nice effect that the 2nd state class (`Builder`) now
only needs to know about the non-correct-guess-letters, and doesn't
need to know about the corresponding 'actual' letters - its code gets
a little simpler as a result.

This is probably slightly over-engineering things, given that for
advanced Wordle analysis the results of `WordFeedback.feedbackFor()`
are all pre-calculated _anyway_, so the speed of the method doesn't
matter too much, and the clarity of the code was already much better
with recent improvements, but I think it pleases me that, of the
computation we're now explcitly doing, there's no obvious repeated
computation (ie. we only check if a guess letter is correct once,
rather than twice!).

I've also added some explanatory scaladoc to help explain what the
two passes are doing.